% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic-spatialstats.R
\name{.local_comp_score}
\alias{.local_comp_score}
\title{Internal: local composition enrichment score}
\usage{
.local_comp_score(frac, n_neighbors, overlap_area, global_density)
}
\value{
Numeric vector of local enrichment scores (percentage-point
difference between observed and expected local composition).
}
\description{
Internal: local composition enrichment score
}
\keyword{internal}
