% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic-extract.R
\name{.panoramic_add_coloc_direction}
\alias{.panoramic_add_coloc_direction}
\title{Internal: add directional colocalization labels}
\usage{
.panoramic_add_coloc_direction(df)
}
\description{
Internal: add directional colocalization labels
}
\keyword{internal}
