% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic-extract.R
\name{.panoramic_extract_meta_table}
\alias{.panoramic_extract_meta_table}
\title{Internal: extract feature-level meta-analysis table from rowData}
\usage{
.panoramic_extract_meta_table(
  se,
  feature_cols = c("ct1", "ct2", "radius_um", "stat"),
  drop_na = FALSE
)
}
\description{
Internal: extract feature-level meta-analysis table from rowData
}
\keyword{internal}
