% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic-extract.R
\name{.panoramic_extract_spatialstats_table}
\alias{.panoramic_extract_spatialstats_table}
\title{Internal: flatten spatialstats assays to a long table}
\usage{
.panoramic_extract_spatialstats_table(se, drop_na = FALSE)
}
\description{
Internal: flatten spatialstats assays to a long table
}
\keyword{internal}
