% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic_prepare.R
\name{.reorder_ccw}
\alias{.reorder_ccw}
\title{Internal: reorder hull vertices counterclockwise}
\usage{
.reorder_ccw(coords)
}
\description{
Internal: reorder hull vertices counterclockwise
}
\keyword{internal}
