% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic_prepare.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{Pipe-compatible null operator}
\usage{
a \%||\% b
}
\arguments{
\item{a, }{b Objects to test; if a is NULL b is returned.}
}
\value{
The non-NULL object.
}
\description{
Returns b if a is NULL, otherwise a.
}
\keyword{internal}
