% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic-spatialstats.R
\name{.panoramic_pairs}
\alias{.panoramic_pairs}
\title{Internal: generate CT pairs present in all/most samples}
\usage{
.panoramic_pairs(
  prep_list,
  cell_type = "cell_type",
  include_self = TRUE,
  min_presence = 1
)
}
\description{
Internal: generate CT pairs present in all/most samples
}
\keyword{internal}
