% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic-spatialstats.R
\name{panoramic}
\alias{panoramic}
\title{One-liner convenience: prepare + spatialstats}
\usage{
panoramic(
  spe_list,
  design,
  cell_type = "cell_type",
  pairs = "auto",
  radii_um,
  stat = "Lcross",
  nsim = 200,
  correction = "translate",
  min_cells = 5L,
  concavity = 50,
  window = "concave",
  keep_boot = FALSE,
  seed = 123,
  BPPARAM = BiocParallel::SerialParam()
)
}
\description{
One-liner convenience: prepare + spatialstats
}
