% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoramic-spatialstats.R
\name{panoramic_spatialstats}
\alias{panoramic_spatialstats}
\title{Compute pairwise spatial stats with bootstrap across samples}
\usage{
panoramic_spatialstats(
  prep,
  pairs = "auto",
  radii_um,
  stat = "Lcross",
  nsim = 200,
  correction = "translate",
  keep_boot = FALSE,
  seed = 123,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{prep}{list from \code{pano_prepare()}}

\item{pairs}{"auto" or data.frame with ct1, ct2}

\item{radii_um}{numeric vector of radii at which to evaluate}

\item{stat}{"Lcross" (default) or "Kcross" / "Lest" / "Kest"}

\item{nsim}{bootstrap iterations}

\item{correction}{spatstat edge correction}

\item{keep_boot}{if TRUE, stores lohboot objects in metadata}

\item{seed}{RNG seed}

\item{BPPARAM}{BiocParallel param}
}
\value{
SummarizedExperiment
}
\description{
Returns a SummarizedExperiment with assays:
\itemize{
\item \code{yi} feature × sample effects
\item \code{vi} feature × sample variances (bootstrap)
}
RowData stores \code{ct1}, \code{ct2}, \code{radius_um}, \code{stat}.
}
